refactor: update Calibrator to use SessionConfig and HardwareManager, improve error handling
- Add todo to include pixelFormat in config

+++ code/cv-system/src/cv_system/config.py
@@ -63,7 +63,7 @@ class CalibrationConfig(BaseModel):
 
     # Four corner points in camera coordinates (y, x)
     # Format: [(y1, x1), (y2, x2), (y3, x3), (y4, x4)]
-    camera_corners: list[Tuple[int, int]] = [
+    camera_corners: list[Tuple[float, float]] = [
         (0, 0),
         (0, 424),
         (512, 0),
@@ -72,7 +72,7 @@ class CalibrationConfig(BaseModel):
 
     # Four corner points in projector coordinates (y, x)
     # Format: [(y1, x1), (y2, x2), (y3, x3), (y4, x4)]
-    projector_corners: list[Tuple[int, int]] = [
+    projector_corners: list[Tuple[float, float]] = [
         (0, 0),
         (0, 1080),
         (1920, 0),
